verification done

+++ src/app/blog.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Blog } from './blog.interface';
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable, map } from 'rxjs';
 
 @Injectable({
@@ -16,15 +16,15 @@ export class BlogService {
 //this function returns observable type blog array
   
   getAllBlogs(): Observable<Blog[]> {
-    return this.httpClient.get<Blog[]>(this.baseUrl);
+    return this.httpClient.get<Blog[]>(this.baseUrl,{ headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
 
   addBlog(Blog: Blog) {
-    return this.httpClient.post(this.baseUrl, Blog);
+    return this.httpClient.post(this.baseUrl, Blog, { headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
 
   deleteBlog(id: string) {
-    return this.httpClient.delete(this.baseUrl + '/' + id);
+    return this.httpClient.delete(this.baseUrl + '/' + id, { headers: new HttpHeaders({'Authorization': `Bearer ${localStorage.getItem("token")}` })});
   }
